+++ src/main/java/crying/tools/blocks/OverHardenedCore.java
@@ -1,0 +1,37 @@
+package crying.tools.blocks;
+
+import crying.tools.Crying;
+import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
+import net.minecraft.block.HeavyCoreBlock;
+import net.minecraft.block.MapColor;
+import net.minecraft.block.enums.NoteBlockInstrument;
+import net.minecraft.block.piston.PistonBehavior;
+import net.minecraft.item.ItemGroups;
+import net.minecraft.item.Items;
+import net.minecraft.registry.RegistryKey;
+import net.minecraft.registry.RegistryKeys;
+import net.minecraft.sound.BlockSoundGroup;
+import net.minecraft.util.Identifier;
+
+public class OverHardenedCore extends HeavyCoreBlock {    
+    public OverHardenedCore() {
+        super(
+            Settings.create()
+            .registryKey(RegistryKey.of(RegistryKeys.BLOCK, Identifier.of(Crying.MOD_ID, "over-hardened_core")))
+            .mapColor(MapColor.IRON_GRAY)
+            .instrument(NoteBlockInstrument.SNARE)
+            .sounds(BlockSoundGroup.HEAVY_CORE)
+            .hardness(900.0F)
+            .pistonBehavior(PistonBehavior.NORMAL)
+            .resistance(3200000F)
+            .requiresTool()
+        );
+        create();
+    }
+
+    public void create() {
+        Crying.registerBlock(this, "over-hardened_core", true);
+      
+        ItemGroupEvents.modifyEntriesEvent(ItemGroups.INGREDIENTS).register((itemGroup) -> itemGroup.addAfter(Items.HEAVY_CORE, this.asItem()));
+    }
+}